Adds a forceRegister db entry and refactors the overview resources into seperate classes
