+++ core/src/main/java/io/aiven/klaw/service/SchemaRegstryControllerService.java
@@ -35,7 +35,6 @@ import org.springframework.stereotype.Service;
 @Service
 @Slf4j
 public class SchemaRegstryControllerService {
-
   @Autowired ManageDatabase manageDatabase;
 
   @Autowired ClusterApiService clusterApiService;
